brandkit-server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
loadenv==0.1.1
MarkupSafe==2.1.3
numpy==1.26.1
opencv-python==4.8.1.78
outcome==1.3.0.post0
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
threadpoolctl==3.2.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
Werkzeug==3.0.1
wsproto==1.2.0
